refactor: moved TargeUnit
